Add status filter
On browsing by files and authors, it is now possible to filter the list
using statuses filter.

The selected statuses filter the file/author list after clicking on the
"Refresh" button. But even if you don't refresh this list, the bug list
WILL BE filtered (not a bug, a feature) depending on the selected
statuses.